Added utils for routes, cleaned code and added a function to validate uuids
