+++ app/routes/users.py
@@ -1,30 +1,23 @@
-from fastapi import APIRouter, Depends, HTTPException
-from sqlmodel import select, Session
-from app.models.users import User, UserCreate
-from app.services.db import get_session
-from uuid6 import uuid7
+from .base import APIRouter, HTTPException, Depends, Session, select, uuid7, User, UserCreate, get_session
+from .utils import check_user_exists, check_uuid
 
 usersRouter = APIRouter(
     prefix="/users",
     tags=["users"],
 )
 
-def check_user_exists(session: Session, email=None, username=None):
-    if email:
-        if session.exec(select(User).where(User.email == email)).first():
-            return "Email already registered"
-    
-    if username:
-        if session.exec(select(User).where(User.username == username)).first():
-            return "Username already taken"
-    
-    return None
-
 @usersRouter.get("/{user_id}")
 def get_users(user_id, session: Session = Depends(get_session)) -> User:
-   user = session.exec(select(User).where(User.id == user_id)).first()
+    
+    if not check_uuid(user_id):
+        raise HTTPException(status_code=400, detail="Invalid UUID format")
+    
+    user = session.exec(select(User).where(User.id == user_id)).first()
    
-   return user
+    if not user:
+       raise HTTPException(status_code=404, detail="User not found")
+   
+    return user
 
 @usersRouter.post("/create")
 def create_user(user: UserCreate, session: Session = Depends(get_session)) -> User:
